pyproject.toml
[project]
name = "magic512bot"
version = "0.1.0"
description = ""
authors = [
    {name = "bassmatt", email = "bassmatt0515@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.4.0",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.32",
    "pytest>=8.3.2",
    "requests>=2.32.3",
    "table2ascii>=1.1.3",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy-utils>=0.41.2",
    "ruff>=0.3.3",
    "mypy>=1.15.0",
    "pytest-asyncio>=0.25.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py311"
fix = true

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "UP", "PL", "RUF"]
ignore = ["PLR0913"]  # Ignore too many arguments for now

[tool.ruff.lint.isort]
known-first-party = ["magic512bot"]
combine-as-imports = true
